Olympic Participation Data

This dataset contains historical participation data from the Summer and Winter Olympics, focusing on male and female athletes from 1900 to 2016.

Usage

olympics_participation

Format

'olympics_participation'

A data frame with A data frame with 271116 rows and 15 variables

id

Unique identifier for each athlete

name

Athlete's name

sex

Athlete's gender

age

Athlete's age at the time of the Olympics

height

Athlete's height in centimeters

weight

Athlete's weight in kilograms

team

The country or team the athlete represented

noc

National Olympic Committee region

games

Name of the Olympic games

year

Year of the Olympics

season

Season of the Olympics

city

Host city

sport

Sport category

event

Specific event

medal

Medal won (Gold, Silver, Bronze, or NA)

Source

https://github.com/rfordatascience/tidytuesday/blob/master/data/2024/2024-08-06/readme.md